requirements.txt
asgiref==3.8.1
diff-match-patch==20230430
Django==5.0.6
django-import-export==4.0.7
et-xmlfile==1.1.0
openpyxl==3.1.3
sqlparse==0.5.0
tablib==3.5.0
typing_extensions==4.12.0
